processing/requirements.txt
azure-functions==1.7.0
joblib==1.0.1
numpy==1.20.3
scikit-learn==0.24.2
scipy==1.6.3
threadpoolctl==2.1.0
